Replaced some u8 with u32

+++ src/item_use.c
@@ -1194,8 +1194,8 @@ static void UsePokevialFieldNo(u8 taskId)
 
 void ItemUseOutOfBattle_Pokevial(u8 taskId)
 {
-    u8 currentDoses = PokevialGetDose();
-    u8 numDigits = CountDigits(currentDoses);
+    u32 currentDoses = PokevialGetDose();
+    u32 numDigits = CountDigits(currentDoses);
 
     CopyItemName(ITEM_POKEVIAL, gStringVar1);
